Extract types into separate files
