+++ Funcky.Analyzers/Funcky.Analyzers.CodeFixes/CompilationExtensions.cs
@@ -1,0 +1,10 @@
+using Microsoft.CodeAnalysis;
+
+namespace Funcky.Analyzers;
+
+internal static class CompilationExtensions
+{
+    public static INamedTypeSymbol? GetGenericOptionType(this Compilation compilation) => compilation.GetTypeByMetadataName("Funcky.Monads.Option`1");
+
+    public static INamedTypeSymbol? GetOptionType(this Compilation compilation) => compilation.GetTypeByMetadataName("Funcky.Monads.Option");
+}
